RealESRGAN/deploy/requirements.txt
numpy
opencv-python
Pillow
fastapi==0.101.0
uvicorn==0.23.2
python-multipart==0.0.6
git+https://github.com/ndominutti/BasicSR.git
